pyproject.toml
[project]
name = "langchain-hello-world"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "dotenv>=0.9.9",
    "langchain>=1.2.13",
    "langchain-google-genai>=4.2.1",
    "google-genai>=0.8.3",
    "langchain-ollama>=1.0.1",
    "langchain-openai>=1.1.12",
]
